+++ infra/binder/src/main/java/org/apache/shardingsphere/infra/binder/engine/type/DALStatementBindEngine.java
@@ -19,6 +19,7 @@ package org.apache.shardingsphere.infra.binder.engine.type;
 
 import lombok.RequiredArgsConstructor;
 import org.apache.shardingsphere.infra.binder.engine.statement.SQLStatementBinderContext;
+import org.apache.shardingsphere.infra.binder.engine.statement.dal.ExplainStatementBinder;
 import org.apache.shardingsphere.infra.binder.engine.statement.dal.OptimizeTableStatementBinder;
 import org.apache.shardingsphere.infra.binder.engine.statement.dal.ShowColumnsStatementBinder;
 import org.apache.shardingsphere.infra.binder.engine.statement.dal.ShowCreateTableStatementBinder;
@@ -26,6 +27,7 @@ import org.apache.shardingsphere.infra.binder.engine.statement.dal.ShowIndexStat
 import org.apache.shardingsphere.infra.hint.HintValueContext;
 import org.apache.shardingsphere.infra.metadata.ShardingSphereMetaData;
 import org.apache.shardingsphere.sql.parser.statement.core.statement.dal.DALStatement;
+import org.apache.shardingsphere.sql.parser.statement.core.statement.dal.ExplainStatement;
 import org.apache.shardingsphere.sql.parser.statement.core.statement.dal.OptimizeTableStatement;
 import org.apache.shardingsphere.sql.parser.statement.core.statement.dal.ShowColumnsStatement;
 import org.apache.shardingsphere.sql.parser.statement.core.statement.dal.ShowCreateTableStatement;
@@ -62,6 +64,9 @@ public final class DALStatementBindEngine {
         }
         if (statement instanceof ShowIndexStatement) {
             return new ShowIndexStatementBinder().bind((ShowIndexStatement) statement, binderContext);
+        }
+        if (statement instanceof ExplainStatement) {
+            return new ExplainStatementBinder(metaData, currentDatabaseName, hintValueContext).bind((ExplainStatement) statement, binderContext);
         }
         return statement;
     }